accounting/2017: Yann transfers the money from his own bank account
When Google paid out the "stipends" for the mentoring during the Google
Summer of Code 2014, there was no Buildroot Association back then, and
Yann received the funds on behalf of the Buildroot project. See:
    https://elinux.org/Buildroot#Accounting

Since then, the Buildroot Association was eventually registered, and
Yann transferred the registering fees to Thomas: 44.00€. This is listed
in the accounting for 2016, dated 2016-04-14.

Now, on 2017-05-15, Yann eventually transferred the money he had on his
account, to the Association's account. In doing so, he referred to the
webpage listed above, and forgot about the 44.00€ he already wired
Thomas.

This means that Yann only accounted for the money he received from
Google, minus the expenses for the thank-you present offered to Niel,
i.e. 423.14-42.10 = 381.04

So, even though thhe actual money held by Yann on behalf of the
Association was only 337.04€, Yann wired 381.04€.

This explains the discrepancy in the numbers.

Yann considers those extra 44€ to be a donation from him to the
Buildroot Association, which nicely resolves the delta.

Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>