bills/templates: cleanup the french template
The table in the french template has weird, huge leading/trailing spaces
around lines in the cells. Adding more than one line to explain a bill
would result in the bill spanning two pages.

Fix that by using manual empty lines, so we can easily override the
leading/trailing space.

This is the way it is already done in the english template, btw.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>