accounting/2018: issue bills for participants to DevDays ELCE`18
The Buildroot Developers Days before ELCE 2018 were held in a shared
Airbnb. Thomas paid for the Airbnb upfront, which amounted to 638.83€.

These Airbnb fees are split as thus:

  - 50% (319.42€) covered by the Association to host the devdays during
    the two days of the meeting,

  - 50% (319.41€) covered by the persons staying at the Airbnb during
    the night, i.e. 106.47 for each of Matt, Romain, and Thomas.

Issue bills for Matt, Romain, and Thomas.

Reimbursing Thomas will be done in a later transaction. Yet, we store
the Airbnb bill now, as a reference for how much Thomas did pay.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>