accounting: fix ledger script when cents are 08 or 09
When the cents are in the range 00 to 07, they are interpreted as octal
numbers, whose values are the correct cecimal value. When the cents are
above 10, they are interpreted as decimal value directly. However, when
they are either 08 or 09, they are still interpreted as octal integers,
but 08 and 09 are invalid octal values.

Rather than doing more complicated computations, we just fix tht with a
bit of shell+sed scriptery: the shell removes the dot '.', and sed
removes any leading zero '0'.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>