accounting/2019: register two memberships and a donation
To be noted: Mat's membership fees have been paid via Paypal, sop the
money is still pending there. We're planning on retrieving money from
there to the bank account in batches, so as to minimize the transfer
fees (if any) from Paypal.

Thus, there is a slight discrepancy between what the bank summary
displays and what the ledger does. But the moeny is there, in the Paypal
account for now.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>