accounting/2019: register GSoC stipends
For our participation to GSoC 2019, GSoC trandferred $3200 stipends for
mentoring a student. At the time, the $3200 turned into €2846.86.

Note: the transfer is from Yann's own account, because he had to rush to
fill in the papers, and in the confusion just forgot the access code to
the Association's own bank account. So, to not lose the stipends, Yann
registered his own bank account with the tranfer intermediary, Payoneer.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>